Revert un-needed changes
