+++ source/common/grpc/BUILD
@@ -81,7 +81,6 @@ envoy_cc_library(
         "//include/envoy/http:header_map_interface",
         "//include/envoy/http:message_interface",
         "//include/envoy/stats:stats_interface",
-        "//include/envoy/stream_info:stream_info_interface",
         "//include/envoy/upstream:cluster_manager_interface",
         "//include/envoy/upstream:upstream_interface",
         "//source/common/buffer:buffer_lib",
