+++ source/common/router/BUILD
@@ -233,7 +233,6 @@ envoy_cc_library(
         "//include/envoy/stats:stats_macros",
         "//include/envoy/upstream:cluster_manager_interface",
         "//include/envoy/upstream:upstream_interface",
-        "//source/common/access_log:access_log_formatter_lib",
         "//source/common/access_log:access_log_lib",
         "//source/common/buffer:watermark_buffer_lib",
         "//source/common/common:assert_lib",
